Classify `CallChain.h` as public API and relocate file
