+++ platform/CallChain.h
@@ -28,11 +28,9 @@ namespace mbed {
 typedef Callback<void()> *pFunctionPointer_t;
 class CallChainLink;
 
-/** \ingroup mbed-os-internal */
-/** \addtogroup platform-internal-api Platform */
-/** @{*/
 /**
  * \defgroup platform_CallChain CallChain class
+ * \ingroup platform-public-api
  * @{
  */
 
@@ -257,8 +255,6 @@ public:
 private:
     CallChainLink *_chain;
 };
-
-/**@}*/
 
 /**@}*/
